feat: Integrate new WorldGenerator in MainScene + Player refactor
