+++ Unity_Minigolf/Assets/Code/Player/PlayerInfo.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: 49f6a89c193241d0889025e8bd77acdb
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
