+++ Unity_Minigolf/Assets/Code/Timer.cs
@@ -1,51 +1,52 @@
 ﻿using System;
-using System.Collections;
-using System.Collections.Generic;
 using UnityEngine;
 using UnityEngine.UI;
 
-public class Timer : MonoBehaviour
+namespace Code
 {
-    private float _timeLimit = 60;
-    private float _remainingTime;
-    public Text timerText;
-
-    private void Start()
+    public class Timer : MonoBehaviour
     {
-        Debug.Log(String.Format("Set time limit to {0}", _timeLimit));
-    }
+        private float _timeLimit = 60;
+        private float _remainingTime;
+        public Text timerText;
 
-    void Update()
-    {
-        _remainingTime = _timeLimit - Time.timeSinceLevelLoad;
-        DisplayTime(_remainingTime);
-        // start displaying the remaining time in red when there are only 10 secs left
-        if (_remainingTime <= 11)
+        private void Start()
         {
-            DisplayTimeInRed(_remainingTime);
+            Debug.Log(String.Format("Set time limit to {0}", _timeLimit));
         }
-        // don't display any negative values
-        if (_remainingTime <= 1)
+
+        void Update()
         {
-            DisplayTimeInRed(0);
+            _remainingTime = _timeLimit - Time.timeSinceLevelLoad;
+            DisplayTime(_remainingTime);
+            // start displaying the remaining time in red when there are only 10 secs left
+            if (_remainingTime <= 11)
+            {
+                DisplayTimeInRed(_remainingTime);
+            }
+            // don't display any negative values
+            if (_remainingTime <= 1)
+            {
+                DisplayTimeInRed(0);
+            }
         }
-    }
     
-    void DisplayTime(float timeToDisplay)
-    {
-        float minutes = Mathf.FloorToInt(timeToDisplay / 60);  
-        float seconds = Mathf.FloorToInt(timeToDisplay % 60);
+        void DisplayTime(float timeToDisplay)
+        {
+            float minutes = Mathf.FloorToInt(timeToDisplay / 60);  
+            float seconds = Mathf.FloorToInt(timeToDisplay % 60);
 
-        timerText.text = string.Format("{0:00}:{1:00}", minutes, seconds);
-    }
+            timerText.text = string.Format("{0:00}:{1:00}", minutes, seconds);
+        }
 
-    void DisplayTimeInRed(float timeToDisplay)
-    {
-        float minutes = Mathf.FloorToInt(timeToDisplay / 60);  
-        float seconds = Mathf.FloorToInt(timeToDisplay % 60);
+        void DisplayTimeInRed(float timeToDisplay)
+        {
+            float minutes = Mathf.FloorToInt(timeToDisplay / 60);  
+            float seconds = Mathf.FloorToInt(timeToDisplay % 60);
 
-        timerText.text = string.Format("{0:00}:{1:00}", minutes, seconds);
-        timerText.color = Color.red;
-    }
+            timerText.text = string.Format("{0:00}:{1:00}", minutes, seconds);
+            timerText.color = Color.red;
+        }
     
+    }
 }
