+++ Unity_Minigolf/Assets/Code/World/ObjectPool.cs
@@ -4,7 +4,7 @@ using System.Linq;
 using UnityEngine;
 using Random = UnityEngine.Random;
 
-namespace Code.WorldGeneration
+namespace Code.World
 {
     [Serializable]
     public class ObjectPoolItem
